+++ README.md
@@ -60,7 +60,7 @@ These plugins help with project development.
 | [SBT PGP][sbt-pgp-link]                     | PGP plugin for `sbt`                                            | BSD 3-Clause License |
 | [SBT Git][sbt-git-link]                     | A plugin for Git integration, used to version the release jars  | BSD 2-Clause License |
 
-[travis-badge]: https://img.shields.io/travis/exasol/import-export-udf-common-scala/master.svg?logo=travis
+[travis-badge]: https://img.shields.io/travis/com/exasol/import-export-udf-common-scala/master.svg?logo=travis
 [travis-link]: https://travis-ci.com/exasol/import-export-udf-common-scala
 [coveralls-badge]: https://coveralls.io/repos/github/exasol/import-export-udf-common-scala/badge.svg?branch=master
 [coveralls-link]: https://coveralls.io/github/exasol/import-export-udf-common-scala?branch=master
