+++ doc/changes/changes_0.2.0.md
@@ -1,9 +1,10 @@
-# Import Export UDF Common Scala 0.1.1, released 2020-10-DD
+# Import Export UDF Common Scala 0.2.0, released 2020-10-DD
 
 ## Features
 
 * #9: Added SLF4J Logging Library as Common Dependency (PR #10)
-* #11: Added Support for Complex Avro Types (Array, Map, Record) (PR #12)
+* #11: Added Support for Avro Complex Types (Array, Map, Record) (PR #12)
+* #13: Added Support for Avro Logical Types (BigDecimal, Date, Timestamp) (PR #14)
 
 ## Documentation
 
@@ -19,10 +20,10 @@
 
 ### Test Dependency Updates
 
-* Updated `org.mockito:mockito-core` from `3.5.10` to `3.5.15`.
+* Updated `org.mockito:mockito-core` from `3.5.10` to `3.6.0`.
 
 ### Plugin Updates
 
 * Updated `com.github.cb372:sbt-explicit-dependencies` from `0.2.13` to `0.2.15`.
-* Updated `org.wartremover:sbt-wartremover` from `2.4.10` to `2.4.11`.
-* Updated `org.wartremover:sbt-wartremover-contib` from `1.3.8` to `1.3.9`.
+* Updated `org.wartremover:sbt-wartremover` from `2.4.10` to `2.4.12`.
+* Updated `org.wartremover:sbt-wartremover-contib` from `1.3.8` to `1.3.10`.
